+++ source/Patches/Enemies.cs
@@ -160,22 +160,20 @@ namespace LethalFixes.Patches
             }
         }
 
-        internal static MethodInfo Internal_GoTowardsEntrance = AccessTools.Method(typeof(MaskedPlayerEnemy), "GoTowardsEntrance");
-        internal static MethodInfo Internal_TeleportMaskedEnemyAndSync = AccessTools.Method(typeof(MaskedPlayerEnemy), "TeleportMaskedEnemyAndSync");
         [HarmonyPatch(typeof(MaskedPlayerEnemy), "DoAIInterval")]
         [HarmonyPostfix]
-        public static void MaskedEnemy_DoAIInterval(MaskedPlayerEnemy __instance, MineshaftElevatorController ___elevatorScript, float ___interestInShipCooldown, float ___timeAtLastUsingEntrance, Vector3 ___mainEntrancePosition)
+        public static void MaskedEnemy_DoAIInterval(MaskedPlayerEnemy __instance)
         {
-            if (GameNetworkManager.Instance.gameVersionNum < 64 && !__instance.isEnemyDead && __instance.currentBehaviourStateIndex == 0 && ___elevatorScript == null)
+            if (GameNetworkManager.Instance.gameVersionNum < 64 && !__instance.isEnemyDead && __instance.currentBehaviourStateIndex == 0 && __instance.elevatorScript == null)
             {
-                if (!(___interestInShipCooldown >= 17f && Vector3.Distance(__instance.transform.position, StartOfRound.Instance.elevatorTransform.position) < 22f))
+                if (!(__instance.interestInShipCooldown >= 17f && Vector3.Distance(__instance.transform.position, StartOfRound.Instance.elevatorTransform.position) < 22f))
                 {
-                    if (Time.realtimeSinceStartup - ___timeAtLastUsingEntrance > 3f && !__instance.GetClosestPlayer(!__instance.isOutside, false, false))
+                    if (Time.realtimeSinceStartup - __instance.timeAtLastUsingEntrance > 3f && !__instance.GetClosestPlayer(!__instance.isOutside, false, false))
                     {
-                        bool flag2 = (bool)Internal_GoTowardsEntrance?.Invoke(__instance, []);
-                        if (Vector3.Distance(__instance.transform.position, ___mainEntrancePosition) < 1f)
+                        bool flag2 = __instance.GoTowardsEntrance();
+                        if (Vector3.Distance(__instance.transform.position, __instance.mainEntrancePosition) < 1f)
                         {
-                            Internal_TeleportMaskedEnemyAndSync?.Invoke(__instance, [RoundManager.FindMainEntrancePosition(true, !__instance.isOutside), !__instance.isOutside]);
+                            __instance.TeleportMaskedEnemyAndSync(RoundManager.FindMainEntrancePosition(true, !__instance.isOutside), !__instance.isOutside);
                         }
                         else if (flag2 && __instance.searchForPlayers.inProgress)
                         {